rearrange config
