+++ src/lksearch/config/__init__.py
@@ -44,7 +44,7 @@ def get_config_file():
 def create_config_file(overwrite: bool = False):
     """Creates a default configuration file in the config directory"""
 
-    from .. import conf
+    from . import config
 
     # check if config file exists
     path_to_config_file = get_config_file()
@@ -80,9 +80,8 @@ def get_cache_dir():
 
     See `~lksearch.Conf` for more information.
     """
-    from .. import conf
 
-    cache_dir = conf.cache_dir
+    cache_dir = config.cache_dir
     if cache_dir is None or cache_dir == "":
         cache_dir = _config.get_cache_dir(ROOTNAME)
     cache_dir = _ensure_cache_dir_exists(cache_dir)
@@ -129,3 +128,92 @@ def clearcache(test=True):
                 print(f"removing {f}")
             else:
                 shutil.rmtree(f)
+
+
+class Config(_config.ConfigNamespace):
+    """
+    Configuration parameters for `search`.
+
+    Refer to `astropy.config.ConfigNamespace` for API details.
+
+    Refer to `Astropy documentation <https://docs.astropy.org/en/stable/config/index.html#accessing-values>`_
+    for usage.
+
+    The attributes listed below are the available configuration parameters.
+
+    Parameters
+    ----------
+    search_result_display_extra_columns
+        List of extra columns to be included when displaying a SearchResult object.
+
+    cache_dir
+        Default cache directory for data files downloaded, etc. Defaults to ``~/.lksearch/cache`` if not specified.
+
+    CLOUD_ONLY
+        If False (default), will download a file whether the source is the cloud or MAST archives.
+        If True, will only download/return uris for data located in the cloud (Amazon S3).
+
+    PREFER_CLOUD
+        Use Cloud-based data product retrieval where available (primarily Amazon S3 buckets for MAST holdings)
+
+    DOWNLOAD_CLOUD
+       Download cloud based data. If False, download() will return a pointer to the cloud based data instead of
+       downloading it - intended usage for cloud-based science platforms (e.g. TIKE)
+
+    CHECK_CACHED_FILE_SIZES
+        Toggles whether to send requests to check whether the size of files in the local cache match the expected online file.
+
+    """
+
+    # Note: when using list or string_list datatype,
+    # the behavior of astropy's parsing of the config file value:
+    # - it does not accept python list literal
+    # - it accepts a comma-separated list of string
+    #   - for a single value, it needs to be ended with a comma
+    # see: https://configobj.readthedocs.io/en/latest/configobj.html#the-config-file-format
+    search_result_display_extra_columns = _config.ConfigItem(
+        [],
+        "List of extra columns to be included when displaying a SearchResult object.",
+        cfgtype="string_list",
+        module="lksearch",
+    )
+
+    cache_dir = _config.ConfigItem(
+        None,
+        "Default cache directory for data files downloaded, etc.",
+        cfgtype="string",
+        module="lksearch.config",
+    )
+
+    CLOUD_ONLY = _config.ConfigItem(
+        False,
+        "Only Download cloud based data."
+        "If False, will download all data"
+        "If True, will only download data located on a cloud (Amazon S3) bucket",
+        cfgtype="boolean",
+    )
+
+    PREFER_CLOUD = _config.ConfigItem(
+        True,
+        "Prefer Cloud-based data product retrieval where available",
+        cfgtype="boolean",
+    )
+
+    DOWNLOAD_CLOUD = _config.ConfigItem(
+        True,
+        "Download cloud based data."
+        "If False, download() will return a pointer to the cloud based data"
+        "instead of downloading it - intended usage for cloud-based science platforms (e.g. TIKE)",
+        cfgtype="boolean",
+    )
+
+    CHECK_CACHED_FILE_SIZES = _config.ConfigItem(
+        True,
+        "Whether to send requests to check the size of files in the cache match the expected online file."
+        "If False, lksearch will assume files within the cache are complete and will not check their file size."
+        "Setting to True will create a modest speed up to retrieving paths for cached files, but will be lest robust, and return an 'UNKNOWN' status message",
+        cfgtype="boolean",
+    )
+
+
+config = Config()
