+++ src/lksearch/kepler.py
@@ -6,14 +6,17 @@ import numpy as np
 from astropy import units as u
 from astropy.coordinates import SkyCoord
 
-from .MASTSearch import MASTSearch
+from .mast import MASTSearch
+from . import PACKAGEDIR, REPR_COLUMNS
 
 pd.options.display.max_rows = 10
 
+REPR_COLUMNS.insert(3, "quarter")
 
-class K2Search(MASTSearch):
+
+class KeplerSearch(MASTSearch):
     """
-    Search Class that queries mast for observations performed by the K2
+    Search Class that queries mast for observations performed by the Kepler
     Mission, and returns the results in a convenient table with options to download.
     By default both mission products and HLSPs are returned.
 
@@ -24,16 +27,16 @@ class K2Search(MASTSearch):
         coordinates in decimal degrees (tuple), or Astropy `~astropy.coordinates.SkyCoord` Object.
     obs_table:Optional[pd.DataFrame] = None
         Optionally, can provice a Astropy `~astropy.table.Table` Object from
-        AstroQuery `~astroquery.mast.Observations.query_criteria` which will be used to construct the observations table
+        AstroQuery `astroquery.mast.Observations.query_criteria` which will be used to construct the observations table
     prod_table:Optional[pd.DataFrame] = None
         Optionally, if you provide an obs_table, you may also provide a products table of assosciated products.  These
         two tables will be concatenated to become the primary joint table of data products.
     table:Optional[pd.DataFrame] = None
-        Optionally, may provide an astropy `~astropy.table.Table` Object  that is the already merged joint table of obs_table
+        Optionally, may provide an stropy `~astropy.table.Table` Object  that is the already merged joint table of obs_table
         and prod_table.
     search_radius:Optional[Union[float,u.Quantity]] = None
         The radius around the target name/location to search for observations.  Can be provided in arcseconds (float) or as an
-        AstroPy `~astropy.units.u.Quantity` Object
+        AstroPy `~astropy.units.Quantity` Object
     exptime:Optional[Union[str, int, tuple]] = (0,9999)
         Exposure time to filter observation results on.  Can be provided as a mission-specific string,
         an int which forces an exact match to the time in seconds, or a tuple, which provides a range to filter on.
@@ -41,15 +44,17 @@ class K2Search(MASTSearch):
         Mission(s) for which to search for data on
     pipeline:  Optional[Union[str, list[str]]] = ["Kepler", "K2", "SPOC"]
         Pipeline(s) which have produced the observed data
-    campaign: Optional[Union[int, list[int]]]  = None,
-        K2 Observing Campaign(s) for which to search for data.
+    quarter: Optional[Union[int, list[int]]] = None,
+        Kepler Observing Quarter(s) for which to search for data.
+    month: Optional[int] = None,
+        Observation month for Kepler
     """
 
     _REPR_COLUMNS = [
         "target_name",
         "pipeline",
         "mission",
-        "campaign",
+        "quarter",
         "exptime",
         "distance",
         "year",
@@ -65,14 +70,16 @@ class K2Search(MASTSearch):
         search_radius: Optional[Union[float, u.Quantity]] = None,
         exptime: Optional[Union[str, int, tuple]] = (0, 9999),
         pipeline: Optional[Union[str, list[str]]] = None,
-        campaign: Optional[Union[int, list[int]]] = None,
+        quarter: Optional[Union[int, list[int]]] = None,
+        month: Optional[int] = None,
         hlsp: bool = True,
     ):
         if hlsp is False:
-            pipeline = ["K2"]
-            self.mission_search = ["K2"]
+            pipeline = ["Kepler"]
+            self.mission_search = ["Kepler"]
         else:
-            self.mission_search = ["K2", "HLSP"]
+            self.mission_search = ["Kepler", "HLSP"]
+
         super().__init__(
             target=target,
             mission=self.mission_search,
@@ -82,21 +89,22 @@ class K2Search(MASTSearch):
             search_radius=search_radius,
             exptime=exptime,
             pipeline=pipeline,
-            sequence=campaign,
+            sequence=None,
         )
-
         if table is None:
-            self._add_K2_mission_product()
-            self._fix_K2_sequence()
-            self._sort_K2()
+            self._add_kepler_mission_product()
+            self._get_sequence_number()
+            self._sort_Kepler()
+            # Can't search mast with quarter/month directly, so filter on that after the fact.
+            self.table = self.table[self._filter_kepler(quarter, month)]
 
     @property
-    def campaign(self):
-        """K2 Observing campaign for each data product found."""
-        return self.table["campaign"].values
+    def quarter(self):
+        """Kepler Observing quarter for each data product found."""
+        return self.table["quarter"].values
 
     @property
-    def HLSPs(self):
+    def hlsps(self):
         """return a MASTSearch object with self.table only containing High Level Science Products"""
         mask = self.table["mission_product"]
         return self._mask(~mask)
@@ -108,43 +116,123 @@ class K2Search(MASTSearch):
         return self._mask(mask)
 
     def _check_exact(self, target):
-        """Was a K2 target ID passed?"""
-        return re.match(r"^(ktwo|epic) ?(\d+)$", target)
+        """Was a Kepler target ID passed?"""
+        return re.match(r"^(kplr|kic) ?(\d+)$", target)
 
     def _target_to_exact_name(self, target):
-        "parse K2 TIC to exact target name"
-        return f"ktwo{target.group(2).zfill(9)}"
+        "parse Kepler TIC to exact target name"
+        return f"kplr{target.group(2).zfill(9)}"
 
     #
-    def _add_K2_mission_product(self):
+    def _handle_kbonus(self):
+        # KBONUS times are masked as they are invalid for the quarter data
+        # kbonus_mask = self.table["pipeline"] == "KBONUS-BKG"
+
+        # self.table['start_time'][kbonus_mask] = something
+        raise NotImplementedError
+
+    def _add_kepler_mission_product(self):
         """Determine whick products are HLSPs and which are mission products"""
         mission_product = np.zeros(len(self.table), dtype=bool)
-        mission_product[self.table["pipeline"] == "K2"] = True
+        mission_product[self.table["pipeline"] == "Kepler"] = True
         self.table["mission_product"] = mission_product
 
-    def _fix_K2_sequence(self):
-        """K2 campaigns 9, 10, and 11 were split into two sections
-        # list these separately in the table with suffixes 'a' and 'b'"""
+    def _get_sequence_number(self):
+        # Kepler sequence_number values were not populated at the time of
+        # writing this code, so we parse them from the description field.
         seq_num = self.table["sequence_number"].values.astype(str)
 
-        mask = self.table["sequence_number"].isin([9, 10, 11])
+        mask = (
+            (self.table["project"] == "Kepler")
+            & self.table["sequence_number"].isna()
+            & ~self.table["description"].str.contains("Data Validation")
+        )
+        re_expr = r".*Q(\d+)"
+        seq_num[mask] = [
+            re.findall(re_expr, item[0])[0] if re.findall(re_expr, item[0]) else ""
+            for item in self.table.loc[mask, ["description"]].values
+        ]
 
-        for index, row in self.table[mask].iterrows():
-            for half, letter in zip([1, 2], ["a", "b"]):
-                if f"c{row['sequence_number']}{half}" in row["productFilename"]:
-                    seq_num[index] = f"{int(row['sequence_number']):02d}{letter}"
+        self.table["sequence_number"] = seq_num
+        seq_num = [int(x) if x != "<NA>" else 99 for x in seq_num]
+        # Create a 'Quarter' column
+        self.table["quarter"] = seq_num
 
-        self.table["campaign"] = seq_num
+        """self.table["mission"] = [
+            f"{proj} - Q{seq}"
+            if seq !=  '<NA>' else f"{proj}" for proj, seq in zip(self.table["mission"].values.astype(str), seq_num)  
+        ]"""
 
-    def _sort_K2(self):
-        # No specific preference for K2 HLSPs
+    def _filter_kepler(
+        self,
+        quarter: Union[int, list[int]] = None,
+        month: Union[int, list[int]] = None,
+    ) -> pd.DataFrame:
+        """Filter Kepler product by month/quarter
+        Returns a boolean mask"""
+        import os
+
+        products = self.table
+
+        mask = np.ones(len(products), dtype=bool)
+
+        if sum(mask) == 0:
+            return products
+
+        # Identify quarter by the description.
+        # This is necessary because the `sequence_number` field was not populated
+        # for Kepler prime data at the time of writing this function.
+        if quarter is not None:
+            quarter_mask = np.zeros(len(products), dtype=bool)
+            for q in np.atleast_1d(quarter).tolist():
+                quarter_mask += products["description"].str.endswith(f"Q{q}")
+            mask &= quarter_mask
+
+        # For Kepler short cadence data the month can be specified
+        if month is not None:
+            month = np.atleast_1d(month).tolist()
+            # Get the short cadence date lookup table.
+            table = pd.read_csv(
+                os.path.join(PACKAGEDIR, "data", "short_cadence_month_lookup.csv")
+            )
+            # The following line is needed for systems where the default integer type
+            # is int32 (e.g. Windows/Appveyor), the column will then be interpreted
+            # as string which makes the test fail.
+            table["StartTime"] = table["StartTime"].astype(str)
+            # Grab the dates of each of the short cadence files.
+            # Make sure every entry has the correct month
+            is_shortcadence = mask & products["description"].str.contains("Short")
+
+            for idx in np.where(is_shortcadence)[0]:
+                quarter = np.atleast_1d(
+                    int(
+                        products["description"][idx]
+                        .split(" - ")[-1]
+                        .replace("-", "")[1:]
+                    )
+                ).tolist()
+                date = (
+                    products["dataURI"][idx].split("/")[-1].split("-")[1].split("_")[0]
+                )
+                # Check if the observation date matches the specified Quarter/month from the lookup table
+                if (
+                    date
+                    not in table["StartTime"][
+                        table["Month"].isin(month) & table["Quarter"].isin(quarter)
+                    ].values
+                ):
+                    mask[idx] = False
+        return mask
+
+    def _sort_Kepler(self):
         sort_priority = {
-            "K2": 1,
+            "Kepler": 1,
+            "KBONUS-BKG": 2,
         }
         df = self.table
         df["sort_order"] = self.table["pipeline"].map(sort_priority).fillna(9)
         df = df.sort_values(
-            by=["distance", "sort_order", "campaign", "pipeline", "exptime"],
+            by=["distance", "sort_order", "quarter", "pipeline", "exptime"],
             ignore_index=True,
         )
         self.table = df
@@ -159,9 +247,10 @@ class K2Search(MASTSearch):
         year: Union[int, list[int], tuple[int]] = None,
         description: Union[str, list[str]] = None,
         filetype: Union[str, list[str]] = None,
-        campaign: Union[int, list] = None,
         limit: int = None,
         inplace=False,
+        quarter: Optional[int] = None,
+        month: Optional[int] = None,
     ):
         """
         Filters the search result table by specified parameters
@@ -186,8 +275,10 @@ class K2Search(MASTSearch):
             a tuple will look for descriptions containing all the keywords.
         filetype : str or list[str]], optional
             Type of product. A list will look for multiple filetypes.
-        campaign : Optional[int], optional
-            K2 observing campaign, by default None
+        quarter : Optional[int], optional
+            Kepler observing quarter, by default None
+        month : Optional[int], optional
+            Kepler observing month, by default None
         limit : int, optional
             how many rows to return, by default None
         inplace : bool, optional
@@ -195,8 +286,9 @@ class K2Search(MASTSearch):
 
         Returns
         -------
-        K2Search object with updated table or None if `inplace==True`
+        KeplerSearch object with updated table or None if `inplace==True`
         """
+        mask = np.ones(len(self.table), dtype=bool)
         mask = self._filter(
             target_name=target_name,
             filetype=filetype,
@@ -206,8 +298,10 @@ class K2Search(MASTSearch):
             description=description,
             pipeline=pipeline,
             mission=mission,
-            sequence_number=campaign,
         )
+
+        if (quarter is not None) | (month is not None):
+            mask = mask & self._filter_kepler(quarter=quarter, month=month)
         if limit is not None:
             cusu = np.cumsum(mask)
             if max(cusu) > limit:
